0lith-desktop/py-backend/requirements.txt
mem0ai>=1.0.0
qdrant-client>=1.7.0
requests>=2.31.0
ollama>=0.4.0
watchdog>=3.0.0
psutil>=5.9.0
